* development:
  Anzeige der LeLe Controls aus, wenn es sich um keine LeLe Phase handelt
  Project Board Customization wird ignoriert, ähnlich wie anderen in ../requirements vorhandene Dateien. ../requirements sollte immer separat mitgeführt werden , also ein Folder für dies Einstellungen, ein anderen für jene ..
  Jira Issue PT-33 Implicit Conversion = off, late binding so gut wie es irgend geht vermieden (bis auf selection.shaperange - hier wurde noch kein Weg gefunden)
  Jira Issue PT-29 (fehlerhafte Powerpoint Charts) bereinigt

+++ Projectboard/ProjectBoardDefinitions/clsPhase.vb
@@ -64,7 +64,7 @@
     End Property
 
 
-    Public Sub changeStartandDauer(ByVal startOffset As Integer, ByVal dauer As Integer)
+    Public Sub changeStartandDauer(ByVal startOffset As Long, ByVal dauer As Long)
 
         Dim projektStartdate As Date
         Dim projektstartColumn As Integer
@@ -134,7 +134,7 @@
                 Catch ex As Exception
 
                 End Try
-                
+
 
                 If awinSettings.autoCorrectBedarfe Then
 
@@ -160,7 +160,7 @@
 
                     End If
 
-                   
+
                 End If
